Restructured project folders for clarity and updated package.json
