+++ integration-tests/api-gateway/crypto/create-jar-request-payload.ts
