+++ integration-tests/api-gateway/e2e/e2e-happy-path.test.ts
@@ -2,15 +2,15 @@ import { getSSMParameter } from "../../step-functions/aws/resources/ssm-param-he
 import {
   Payload,
   getJarAuthorizationPayload,
-} from "./crypto/create-jar-request-payload";
+} from "../crypto/create-jar-request-payload";
 import {
   NINO,
   CLIENT_ID,
   getClaimSet,
   CLIENT_URL,
   environment,
-} from "./env-variables";
-import { buildPrivateKeyJwtParams } from "./crypto/client";
+} from "../env-variables";
+import { buildPrivateKeyJwtParams } from "../crypto/client";
 import { JWK } from "jose";
 import {
   clearAttemptsTable,
