+++ integration-tests/api-gateway/jest.config.ts
@@ -1,8 +1,8 @@
 import type { Config } from "jest";
-import baseConfig from "../../jest.config";
+import baseConfig from "../jest.config";
 
 export default {
   ...baseConfig,
   projects: [],
-  displayName: "integration-tests/api-gateway/aws",
+  displayName: "integration-tests/api-gateway",
 } satisfies Config;
