+++ integration-tests/package.json
@@ -6,8 +6,8 @@
     "unit:all": "jest --runInBand",
     "unit:aws": "jest --silent --select-projects integration-tests/aws --runInBand",
     "unit:mocked": "jest --silent --select-projects integration-tests/mocked",
-    "integration:all": "jest --select-projects integration-tests/api-gateway/aws",
-    "integration:e2e": "jest integration-tests/api-gateway/aws/e2e-retry-path.test.ts integration-tests/api-gateway/aws/e2e-happy-path.test.ts",
+    "integration:all": "jest --silent --select-projects integration-tests/api-gateway",
+    "integration:e2e": "jest --silent integration-tests/api-gateway/e2e",
     "test:all": "npm run unit:all --",
     "test:aws": "npm run unit:aws --",
     "test:integration": "npm run integration:all --",
